Create split.py
Merges the data and the targets in the 3 .mat files - um, umm and uu. It then splits them into training, validation and test data and targets in the ratio 60:10:30 respectively.

split_main is the main function that is to be called. It returns the tuple - (training data, training targets, validation data, validation targets, test data, test targets).